MVP-6356: modulize cypress commands
